backend/pyproject.toml
[project]
name = "stockroom"
version = "0.7.1"
description = "The project is designed to maintain a database with equipment and consumables"
authors = [
    {name = "Makarovskiy Nikolay", email = "adrameleh666@gmail.com"},
]

dependencies = [
    "asgiref==3.7.2",
    "attrs==23.1.0",
    "diff-match-patch==20230430",
    "Django >=4.1.6,<=5",
    "djangorestframework==3.14.0",
    "djangorestframework-simplejwt>=5.2.2",
    "markdown==3.4.3",
    "django-filter==23.2",
    "gunicorn==22.0.0",
    "celery==5.4.0",
    "django-celery-beat==2.5.0",
    "redis==4.6.0",
    "hiredis==2.2.3",
    "flower==2.0.0",
    "django-adaptors==0.2.5",
    "django-appconf==1.0.5",
    "django-cors-headers==4.2.0",
    "django-crispy-forms==2.0",
    "django-bootstrap5==23.3",
    "crispy-bootstrap5==0.7",
    "Pillow==10.2.0",
    "psycopg2==2.9.6",
    "psycopg2-binary>=2.9.9",
    "python-dateutil==2.8.2",
    "pytz==2023.3",
    "six==1.16.0",
    "sqlparse==0.5.0",
    "django-select2==8.1.2",
    "django-import-export==4.0.6",
]
requires-python = ">=3.12"
readme = "README.md"
license = {file = "LICENSE"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
lint = [
    "ruff==0.4.8",
    "mypy==1.10.0",
    "pre-commit==3.7.1",
    "isort==5.13.2",
    "setuptools>=71.0.1",
    "django-stubs>=5.0.0",
    "djangorestframework-stubs>=3.15.0"
]
test = [
    "pytest==8.1.1",
    "pytest-django==4.8.0",
    "pytest-xdist==3.6.1",
    "pytest-cov==5.0.0",
    "django-debug-toolbar==4.1.0",

    ]

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "__pycache__",
]


[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 4
skip_gitignore = true
skip_glob = [
    "**/migrations/*",
    "**/settings/*",
]
src_paths = [
    "./src",
]

[tool.mypy]
mypy_path = "./src"
follow_imports = "normal"
disallow_any_generics = true
check_untyped_defs = true
no_implicit_reexport = true
show_error_codes = true
warn_unused_ignores = false
show_error_context = true
plugins = ["mypy_django_plugin.main",  "mypy_drf_plugin.main"]

[mypy-pyproj]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "import_export.*",
    "rest_framework.*",
    ]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]
pythonpath = "./src"

[tool.django-stubs]
django_settings_module = "backend.settings.production"
